Code Re-Structure
Peripheral (LO's, attenuators, etc.) are setup in the "instrument_init.py" file
Experimental parameters such as LO frequencies and mixer calibration parameters are stored in a JSON file format, created by "parameter_dict.py".
The configuration file looks into that file to find the values of the parameters it needs.